+++ components/locale/src/canonicalizer.rs
@@ -35,7 +35,7 @@ use tinystr::TinyAsciiStr;
 /// assert_eq!(locale, "ja-Latn-alalc97-fonipa".parse().unwrap());
 /// ```
 ///
-/// [UTS #35: Annex C, LocaleId Canonicalization]: http://unicode.org/reports/tr35/#LocaleId_Canonicalization
+/// [UTS #35: Annex C, LocaleId Canonicalization]: https://unicode.org/reports/tr35/#LocaleId_Canonicalization
 #[derive(Debug)]
 pub struct LocaleCanonicalizer<Expander = LocaleExpander> {
     /// Data to support canonicalization.
@@ -301,7 +301,7 @@ impl<Expander: AsRef<LocaleExpander>> LocaleCanonicalizer<Expander> {
 
     /// The canonicalize method potentially updates a passed in locale in place
     /// depending up the results of running the canonicalization algorithm
-    /// from <http://unicode.org/reports/tr35/#LocaleId_Canonicalization>.
+    /// from <https://unicode.org/reports/tr35/#LocaleId_Canonicalization>.
     ///
     /// Some BCP47 canonicalization data is not part of the CLDR json package. Because
     /// of this, some canonicalizations are not performed, e.g. the canonicalization of